Add files via upload
In the second part of Part 09, I learned more about polymorphic objects within Java and how they can inherit characteristics of the parent object. For example, a String type is automatically also an Object type because a String is a type of Object but not vice versa. I also had more practice with abstract classes, inheritances, and interfaces.