Add files via upload
Part 12 mainly focuses on Java Generics (which about killed me) and how to implement generic lists, arrays, and HashMaps. There are also exercises pertaining to multidimensional arrays and how to implement nested loops to access all the variables in these arrays. I'm stuck on the last exercise figuring out how to do this for a magic square :(.