Add files via upload
In the beginning-middle of Part 13, I learned about GUI in JavaFX. Much of the layout is familiar to me because I'm familiar with CSS and Bootstrap, but it's a good refresher on how to append components to different types of layouts, such as grids, BoderPanes, HBoxes, VBoxes, etc. 

I also created a NotePad clone which displays a word count, character count, and "longest word" components. Similar to the document object model in JS, I would just grab the object using an event handler and update the object once the user begins typing.