Add files via upload
The final part of Part 13 contain exercises dealing with programs that have multiple scenes or states. For example, the dictionary program has two states: the initial state is where users can input words and definitions into the "dictionary" object, and the second state is where users can practice their vocabulary; here, the dictionary object will pull up a random word, asking the users to put in the word's definition. The program then checks the user's input against the translation in the dictionary object and displays feedback on whether the input is correct or incorrect.

The final project is a Tic-Tac-Toe game (which took me a long time to figure out) which combines all the essential elements learned from this chapter: the Tic-Tac-Toe board is made of button objects with text-input, made into a GridPane. The rules of the game are implemented using simple if-else statements which change the text of the button objects.